+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 5. "README.md" in English version
 6. Better API Request page
 7. more comprehensive README and TODO
+8. feature: Account(include server and client)